requirements.txt
llama-cpp-python
fastapi
uvicorn 
streamlit 
langchain
langchain-community
faiss-cpu
sentence-transformers
transformers
requests
beautifulsoup4
python-multipart
pydantic
weaviate-client
PyPDF2
python-docx 